Add files via upload
Add new user implemented

+++ login/src/main/java/io/muic/ooc/webapp/servlet/DeleteUserServlet.java
@@ -9,6 +9,7 @@ import io.muic.ooc.webapp.Routable;
 import io.muic.ooc.webapp.model.User;
 import io.muic.ooc.webapp.service.SecurityService;
 import io.muic.ooc.webapp.service.UserService;
+import org.apache.commons.lang.StringUtils;
 
 import javax.servlet.RequestDispatcher;
 import javax.servlet.ServletException;
@@ -44,15 +45,25 @@ public class DeleteUserServlet extends HttpServlet implements Routable {
             UserService userService = UserService.getInstance();
 
             try {
+                User currentUser = userService.findByUsername(username);
+
                 User deletingUser = userService.findByUsername(request.getParameter("username"));
-                if (userService.deleteUserByUsername(deletingUser.getUsername())) {
-                    // goto success message
-                    request.getSession().setAttribute("hasError", false);
-                    request.getSession().setAttribute("message", String.format("User %s is Successfully Deleted", deletingUser.getUsername()));
-                } else {
+
+                if (StringUtils.equals(currentUser.getUsername(), deletingUser.getUsername())) {
                     request.getSession().setAttribute("hasError", true);
-                    request.getSession().setAttribute("message", String.format("Unable to Delete User %s", deletingUser.getUsername()));
+                    request.getSession().setAttribute("message", "Cannot delete your own account");
                 }
+                else {
+                    if (userService.deleteUserByUsername(deletingUser.getUsername())) {
+                        // goto success message
+                        request.getSession().setAttribute("hasError", false);
+                        request.getSession().setAttribute("message", String.format("User %s is Successfully Deleted", deletingUser.getUsername()));
+                    } else {
+                        request.getSession().setAttribute("hasError", true);
+                        request.getSession().setAttribute("message", String.format("Unable to Delete User %s", deletingUser.getUsername()));
+                    }
+                }
+
             } catch (Exception e) {
                 request.getSession().setAttribute("hasError", true);
                 request.getSession().setAttribute("message", String.format("Unable to Delete User %s", request.getParameter("username")));
